requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
loguru==0.7.2